polish: doc scrubs, contract corrections, small guards (code review §5)
- linalg/dot.go: dotF32's panic said "encoder:" (every other panic says
  "linalg:") and a comment named a private repo ("ken's matmul shapes").
  Corrected the prefix and scrubbed the name.
- linalg/doc.go: "the parallel and blocked paths are bit-identical to the serial
  scalar reference … exact equality" overstated — the exact-equality tests pin
  parallel==serial (same kernel) + width/M-invariance; vs a naive scalar
  triple-loop the blocked kernel is only tolerance-checked (~1e-3). Reworded.
- linalg/weightmat.go: MatmulBT and MatmulBTInto tested their dispatch cases in
  different order (harmless for constructor-built values, mutually exclusive,
  but a hand-built both-set WeightMat would route the two entries to different
  kernels). Aligned the order.
- topk/topk.go: package doc cited ADR-025/026, "ken is on Go 1.26", and
  internal/ann.Flat.Query / internal/bm25.Index.TopK — none exist in this
  module. Scrubbed to the real aikit/ann, aikit/bm25.
- encoder/model.go: documented that SetMaxSeqLength mutates unsynchronized state
  (call before sharing the Model across goroutines) — the type otherwise
  advertises immutable-after-Load.
- encoder/crossencoder.go: LoadCrossEncoder now rejects a degenerate classifier
  (labels < 1 → all[0] panic) and a tokenizer missing [CLS]/[SEP] (pairIDs would
  silently frame the pair with id 0 = [PAD]).
- encoder/quant.go: noted quantizeRowsInt8/dequantizeRowsInt8 are test-only
  references now (production quantizes via linalg.QuantizeInt8) — a drift hazard
  to keep in lock-step.
- vision/preprocess.go: the decompression-bomb guard multiplied two
  attacker-controlled header ints in `int`; on a 32-bit build 65535×65535 wraps
  negative and bypasses it. Compute in int64.
- Fixed doc-comment ordering for the new Close() methods (BERT/CrossEncoder/
  SPLADE) so the following method's godoc isn't absorbed (ST1020).

gofmt/vet/golangci-lint(0)/go test ./... green.

+++ vision/preprocess.go
@@ -65,7 +65,10 @@ func Preprocess(data []byte, cfg Config) (*PixelValues, error) {
 	if ic.Width <= 0 || ic.Height <= 0 {
 		return nil, fmt.Errorf("vision: non-positive image dims %dx%d", ic.Width, ic.Height)
 	}
-	if cfg.MaxPixels > 0 && ic.Width*ic.Height > cfg.MaxPixels {
+	// int64: Width and Height are attacker-controlled header ints, so on a
+	// 32-bit build (386/arm) Width*Height in int can wrap negative and bypass
+	// this decompression-bomb guard.
+	if cfg.MaxPixels > 0 && int64(ic.Width)*int64(ic.Height) > int64(cfg.MaxPixels) {
 		return nil, fmt.Errorf("vision: image %dx%d exceeds %d-pixel limit (decompression bomb?)", ic.Width, ic.Height, cfg.MaxPixels)
 	}
